Optimized inventory to deactivate when not visible
Inventory now deactivates itself when no longer visible, similar to
dialogue. Only complex UI objects will be subject to this. Simple,
single-objects like the Clock or Money will not be subject to this.

However, as a result of this change, a few other things had to be changed
as well. Included are:

- Global UI variables are now prefixed with "ui_". This is to prevent
  namespace clashes with other functions, like global.money or
  global.dialogue.
- Compacted inventory_get_item_slot_*() into one script, merging the _raw
  functions into tabs.
- Inventory is now opened in obj_player, but is still closed in
  obj_ui_inventory. This is to reduce dependancies on other objects.