Moved player variables into player_init()
Player variables that are persistent are now stored in the global
namespace. Only variables pertaining to the instance are stored in
obj_player. For example, the player's name is now stored in
global.player_name rather than the result of player_get_name().

The intro now displays a simple text to choose what to do. This will be
expanded soon.