Fixed clock not showing properly
obj_ui_clock now has a variable "can_show", which functions similar to
global.player_input. It gets set to true/false when it can be shown,
rather than manually checking if the dialogue or inventory is visible.
More decoupling! Or at least the best you can get in GameMaker.