Added mutext_draw
Markup text can now be drawn. It should be noted that it is extremely
inefficient, and drawing it every frame will cause noticable slowdown. The
current solution to this is to draw once to a surface, and display that.
The only optimization technique I see at the moment is to compare integer
values instead of strings. That alone could save a lot of processing.

- Fixed bug with consecutive tags.
- Text with no strings will no longer be added. For example, consecutive
  tags won't have an empty string between them anymore.